Move bank related events to `BankOverlay`
